Added an action creator utility for generic updates.
For these actions state classes need not provide specific reducers,
which reduces the amount of boilerplate code.

State.createUpdateAction(...) is intended for subclasses of State
only (i.e., it should be "protected" if JS supported that).  It should
not be used directly by client code (i.e., UI callbacks) to respect the
encapsulation of the subclass's API.